+++ packages/spectra/source/supports.ts
@@ -1,5 +1,5 @@
 import type {
-    ISpectra,
+    ISpectraElement,
     ISpectraAttributes, ISpectraAttributesBlueprint,
     ISpectraBlueprint,
     ISpectraChildren,
@@ -8,23 +8,21 @@ import type {
     ISpectraStyleKeys, ISpectraStyleValue,
     ISpectraStyleValues
 } from "./types.js";
-import {safeString} from "@protorians/core";
+import {camelCase, safeString, unCamelCase} from "@protorians/core";
 
-export class Spectra implements ISpectra {
+
+export class SpectraElement implements ISpectraElement {
 
     private _blueprint: ISpectraBlueprint = {} as ISpectraBlueprint;
     private _tree: ISpectraRawChildren[] = [];
+    private _removed: boolean = true;
 
     constructor(
-        private readonly name: string,
+        public readonly tagName: string,
     ) {
         this._blueprint.classList = new Set<string>();
         this._blueprint.attributes = new Map<string, string | null | undefined>()
         this._blueprint.style = new Map<ISpectraStyleKeys, ISpectraStyleValue>()
-    }
-
-    get tagName(): string {
-        return this.name;
     }
 
     get blueprint(): ISpectraBlueprint {
@@ -35,12 +33,59 @@ export class Spectra implements ISpectra {
         return this._tree;
     }
 
+    get removed(): boolean {
+        return this._removed;
+    }
+
+    get attributes(): ISpectraAttributes {
+        const attribs = {}
+        for (let [key, value] of this._blueprint.attributes.entries()) {
+            if (!key.startsWith('data-')) attribs[unCamelCase(key)] = value;
+        }
+        return attribs;
+    }
+
+    get dataset(): ISpectraAttributes {
+        const dataset = {}
+        for (let [key, value] of this._blueprint.attributes.entries()) {
+            if (key.startsWith('data-'))
+                dataset[unCamelCase(key).substring('data-'.length)] = value;
+        }
+        return dataset;
+    }
+
+    get textContent(): string {
+        return this._tree
+            .map(child => child instanceof SpectraElement ? child.render() : child)
+            .filter(child => typeof child == 'string')
+            .join('')
+    }
+
+    set textContent(children: string | null | undefined ) {
+        this.append(children)
+    }
+
+    get value(): string {
+        return this.textContent;
+    }
+
+    set value(children: string | null | undefined ) {
+        this.children(children)
+    }
+
+    children(children: ISpectraChildren): this {
+        this._tree = [];
+        this.append(children);
+        return this;
+    }
+
     classname(classname: string | string[]): this {
         (Array.isArray(classname)
             ? classname
             : [...classname.split(' ')])
             .forEach(item => {
-                if (!this._blueprint.classList.has(item))
+                item = item.trim();
+                if (item.length && !this._blueprint.classList.has(item))
                     this._blueprint.classList.add(item)
             });
         return this;
@@ -55,25 +100,48 @@ export class Spectra implements ISpectra {
         return this;
     }
 
-    attributes(attributes: ISpectraAttributes): this {
+    attribute(attributes: ISpectraAttributes): this {
         for (const [key, value] of Object.entries(attributes)) {
-            this._blueprint.attributes.set(key, value);
+            if (!key.startsWith('data-')) {
+                if (typeof value !== 'undefined') this._blueprint.attributes.set(key, value?.toString());
+                else this._blueprint.attributes.delete(key);
+            }
         }
         return this
     }
 
-    text(children: ISpectraChildren): this {
-        const isPromise = children instanceof Promise
-        if (!isPromise) this._tree.push(children)
-        if (isPromise)
-            children.then((value: any) => this.text(value));
+    data(dataset: ISpectraAttributes): this {
+        for (let [key, value] of Object.entries(dataset)) {
+            key = `data-${camelCase(key)}`
+            if (typeof value !== 'undefined') this._blueprint.attributes.set(key, value?.toString());
+            else this._blueprint.attributes.delete(key);
+        }
         return this
     }
 
-    children(children: ISpectraChildren): this {
-        this._tree = [];
-        this.text(children);
-        return this;
+    prepend(children: ISpectraChildren): this {
+        const isPromise = children instanceof Promise
+        if (!isPromise) this._tree.unshift(children)
+        if (isPromise)
+            children.then((value: any) => this.prepend(value));
+        return this
+    }
+
+    append(children: ISpectraChildren): this {
+        const isPromise = children instanceof Promise
+        if (!isPromise) this._tree.push(children)
+        if (isPromise)
+            children.then((value: any) => this.append(value));
+        return this
+    }
+
+    appendChild(child: ISpectraElement): this {
+        this._tree.push(child)
+        return this
+    }
+
+    remove(): void {
+        this._removed = false;
     }
 
     async render(): Promise<string> {
@@ -82,8 +150,14 @@ export class Spectra implements ISpectra {
 }
 
 
-export async function spectraRender(instance: ISpectra): Promise<string> {
+export async function spectraRender(instance: ISpectraElement): Promise<string> {
     return new Promise<string>(async (resolve) => {
+
+        if (!instance.removed) {
+            resolve('')
+            return;
+        }
+
         const sequences: string[] = [];
         const attributes = await spectraAttributesRender(instance.blueprint.attributes);
         const classname = await spectraClassnameRender(instance.blueprint.classList)
@@ -107,7 +181,7 @@ export async function spectraAttributesRender(attributes: ISpectraAttributesBlue
         resolve(
             [...attributes.entries()].map(
                 ([key, value]) =>
-                    value ? `${key}="${safeString(value.toString())}"` : ''
+                    value ? `${unCamelCase(key)}="${safeString(value.toString())}"` : ''
             ).join(" ")
         );
     })
@@ -156,7 +230,7 @@ export async function spectraChildrenRender(children: ISpectraChildren): Promise
         } else if (Array.isArray(children)) {
             for (const child of children)
                 sequences.push(await spectraChildrenRender(child))
-        } else if (children instanceof Spectra) {
+        } else if (children instanceof SpectraElement) {
             sequences.push(await spectraRender(children))
         } else if (typeof children !== 'object') {
             sequences.push(children?.toString() || "")